move step processor in convolutions
